+++ NEWS.md
@@ -1,6 +1,7 @@
 # brandseyer2 0.0.1.9012
 
 * Now has a `with_account()` verb for the filter language.
+* Added a way to nicely see all brands in a query: `get_query_brands()`.
 
 # brandseyer2 0.0.1.9011